# compact: add point tombstone density compaction heuristic