--- a/replay/testdata/replay
+++ b/replay/testdata/replay
@@ -12,7 +12,7 @@ tree
        0      LOCK
       98      MANIFEST-000001
      122      MANIFEST-000008
-    1240      OPTIONS-000003
+    1359      OPTIONS-000003
        0      marker.format-version.000001.013
        0      marker.manifest.000002.MANIFEST-000008
             simple/
@@ -23,7 +23,7 @@ tree
       25        000004.log
      586        000005.sst
       98        MANIFEST-000001
-    1240        OPTIONS-000003
+    1359        OPTIONS-000003
        0        marker.format-version.000001.013
        0        marker.manifest.000001.MANIFEST-000001
 
@@ -60,6 +60,9 @@ cat build/OPTIONS-000003
   multilevel_compaction_heuristic=wamp(0.00, false)
   read_compaction_rate=16000
   read_sampling_multiplier=16
+  num_deletions_threshold=100
+  deletion_size_ratio_threshold=0.500000
+  tombstone_dense_compaction_threshold=0.050000
   strict_wal_tail=true
   table_cache_shards=2
   validate_on_ingest=false