--- a/replay/testdata/replay_paced
+++ b/replay/testdata/replay_paced
@@ -14,7 +14,7 @@ tree
        0      LOCK
      122      MANIFEST-000008
      205      MANIFEST-000011
-    1240      OPTIONS-000003
+    1359      OPTIONS-000003
        0      marker.format-version.000001.013
        0      marker.manifest.000003.MANIFEST-000011
             high_read_amp/
@@ -26,7 +26,7 @@ tree
       39        000009.log
      560        000010.sst
      157        MANIFEST-000011
-    1240        OPTIONS-000003
+    1359        OPTIONS-000003
        0        marker.format-version.000001.013
        0        marker.manifest.000001.MANIFEST-000011