@anish-shanbhag 2024-07-25
I've tested this new heuristic using the benchmarks at #3791 , #3744, and a custom CockroachDB workload. Here are those results.

### `BenchmarkPointDeletedSwath`
<details>
<summary>Show Results</summary>
<pre>
goos: darwin
goarch: arm64
pkg: github.com/cockroachdb/pebble
                                                        │       swath        │              swath-heur              │
                                                        │       sec/op       │    sec/op     vs base                │
PointDeletedSwath/gap=100/prefix_point_lookup-10                7.313µ ± ∞ ¹   7.317µ ± ∞ ¹         ~ (p=0.886 n=4)
PointDeletedSwath/gap=100/non-prefix_point_seek-10              8.923µ ± ∞ ¹   9.834µ ± ∞ ¹         ~ (p=0.057 n=4)
PointDeletedSwath/gap=100/full_scan-10                           2.161 ± ∞ ¹    2.069 ± ∞ ¹    -4.27% (p=0.029 n=4)
PointDeletedSwath/gap=1000/prefix_point_lookup-10               7.465µ ± ∞ ¹   7.141µ ± ∞ ¹         ~ (p=0.343 n=4)
PointDeletedSwath/gap=1000/non-prefix_point_seek-10            11.034µ ± ∞ ¹   8.925µ ± ∞ ¹   -19.11% (p=0.029 n=4)
PointDeletedSwath/gap=1000/full_scan-10                          2.130 ± ∞ ¹    2.095 ± ∞ ¹         ~ (p=0.486 n=4)
PointDeletedSwath/gap=10000/prefix_point_lookup-10              8.123µ ± ∞ ¹   7.676µ ± ∞ ¹         ~ (p=0.114 n=4)
PointDeletedSwath/gap=10000/non-prefix_point_seek-10           10.750µ ± ∞ ¹   9.524µ ± ∞ ¹         ~ (p=0.057 n=4)
PointDeletedSwath/gap=10000/full_scan-10                         2.121 ± ∞ ¹    2.040 ± ∞ ¹         ~ (p=0.343 n=4)
PointDeletedSwath/gap=100000/prefix_point_lookup-10             7.293µ ± ∞ ¹   7.156µ ± ∞ ¹         ~ (p=0.486 n=4)
PointDeletedSwath/gap=100000/non-prefix_point_seek-10          93.112µ ± ∞ ¹   8.794µ ± ∞ ¹   -90.56% (p=0.029 n=4)
PointDeletedSwath/gap=100000/full_scan-10                        2.104 ± ∞ ¹    2.023 ± ∞ ¹    -3.82% (p=0.029 n=4)
PointDeletedSwath/gap=200000/prefix_point_lookup-10             7.321µ ± ∞ ¹   6.960µ ± ∞ ¹    -4.93% (p=0.029 n=4)
PointDeletedSwath/gap=200000/non-prefix_point_seek-10         384.532µ ± ∞ ¹   8.833µ ± ∞ ¹   -97.70% (p=0.029 n=4)
PointDeletedSwath/gap=200000/full_scan-10                        2.113 ± ∞ ¹    2.023 ± ∞ ¹    -4.27% (p=0.029 n=4)
PointDeletedSwath/gap=400000/prefix_point_lookup-10             7.741µ ± ∞ ¹   6.942µ ± ∞ ¹   -10.32% (p=0.029 n=4)
PointDeletedSwath/gap=400000/non-prefix_point_seek-10        1444.402µ ± ∞ ¹   8.915µ ± ∞ ¹   -99.38% (p=0.029 n=4)
PointDeletedSwath/gap=400000/full_scan-10                        2.189 ± ∞ ¹    2.003 ± ∞ ¹    -8.46% (p=0.029 n=4)
PointDeletedSwath/gap=5000000/prefix_point_lookup-10            8.303µ ± ∞ ¹   4.889µ ± ∞ ¹   -41.13% (p=0.029 n=4)
PointDeletedSwath/gap=5000000/non-prefix_point_seek-10     159481.501µ ± ∞ ¹   5.599µ ± ∞ ¹  -100.00% (p=0.029 n=4)
PointDeletedSwath/gap=5000000/full_scan-10                       2.013 ± ∞ ¹    1.125 ± ∞ ¹   -44.09% (p=0.029 n=4)
PointDeletedSwath/gap=10000000/prefix_point_lookup-10           7.694µ ± ∞ ¹   1.944µ ± ∞ ¹   -74.73% (p=0.029 n=4)
PointDeletedSwath/gap=10000000/non-prefix_point_seek-10   1175342.167µ ± ∞ ¹   4.272µ ± ∞ ¹  -100.00% (p=0.029 n=4)
PointDeletedSwath/gap=10000000/full_scan-10                    1788.7m ± ∞ ¹   382.8m ± ∞ ¹   -78.60% (p=0.029 n=4)
geomean                                                         2.076m         411.9µ         -80.16%
¹ need >= 6 samples for confidence interval at level 0.95
                                                        │      swath      │              swath-heur              │
                                                        │      B/op       │     B/op       vs base               │
PointDeletedSwath/gap=100/prefix_point_lookup-10              76.00 ± ∞ ¹     77.00 ± ∞ ¹   +1.32% (p=0.029 n=4)
PointDeletedSwath/gap=100/non-prefix_point_seek-10            76.50 ± ∞ ¹     78.00 ± ∞ ¹   +1.96% (p=0.029 n=4)
PointDeletedSwath/gap=100/full_scan-10                      613.8Ki ± ∞ ¹   812.9Ki ± ∞ ¹  +32.43% (p=0.029 n=4)
PointDeletedSwath/gap=1000/prefix_point_lookup-10             75.50 ± ∞ ¹     77.00 ± ∞ ¹   +1.99% (p=0.029 n=4)
PointDeletedSwath/gap=1000/non-prefix_point_seek-10           77.00 ± ∞ ¹     78.00 ± ∞ ¹   +1.30% (p=0.029 n=4)
PointDeletedSwath/gap=1000/full_scan-10                     613.5Ki ± ∞ ¹   812.7Ki ± ∞ ¹  +32.48% (p=0.029 n=4)
PointDeletedSwath/gap=10000/prefix_point_lookup-10            76.00 ± ∞ ¹     77.00 ± ∞ ¹   +1.32% (p=0.029 n=4)
PointDeletedSwath/gap=10000/non-prefix_point_seek-10          77.00 ± ∞ ¹     78.00 ± ∞ ¹   +1.30% (p=0.029 n=4)
PointDeletedSwath/gap=10000/full_scan-10                    614.0Ki ± ∞ ¹   805.8Ki ± ∞ ¹  +31.23% (p=0.029 n=4)
PointDeletedSwath/gap=100000/prefix_point_lookup-10           76.00 ± ∞ ¹     77.00 ± ∞ ¹   +1.32% (p=0.029 n=4)
PointDeletedSwath/gap=100000/non-prefix_point_seek-10        115.50 ± ∞ ¹     78.00 ± ∞ ¹  -32.47% (p=0.029 n=4)
PointDeletedSwath/gap=100000/full_scan-10                   621.4Ki ± ∞ ¹   802.6Ki ± ∞ ¹  +29.15% (p=0.029 n=4)
PointDeletedSwath/gap=200000/prefix_point_lookup-10           76.00 ± ∞ ¹     77.00 ± ∞ ¹   +1.32% (p=0.029 n=4)
PointDeletedSwath/gap=200000/non-prefix_point_seek-10        138.00 ± ∞ ¹     78.00 ± ∞ ¹  -43.48% (p=0.029 n=4)
PointDeletedSwath/gap=200000/full_scan-10                   627.1Ki ± ∞ ¹   797.6Ki ± ∞ ¹  +27.19% (p=0.029 n=4)
PointDeletedSwath/gap=400000/prefix_point_lookup-10           76.00 ± ∞ ¹     76.50 ± ∞ ¹        ~ (p=0.429 n=4)
PointDeletedSwath/gap=400000/non-prefix_point_seek-10        238.50 ± ∞ ¹     78.00 ± ∞ ¹  -67.30% (p=0.029 n=4)
PointDeletedSwath/gap=400000/full_scan-10                   639.2Ki ± ∞ ¹   790.4Ki ± ∞ ¹  +23.65% (p=0.029 n=4)
PointDeletedSwath/gap=5000000/prefix_point_lookup-10          76.00 ± ∞ ¹     50.00 ± ∞ ¹  -34.21% (p=0.029 n=4)
PointDeletedSwath/gap=5000000/non-prefix_point_seek-10      7340.50 ± ∞ ¹     50.00 ± ∞ ¹  -99.32% (p=0.029 n=4)
PointDeletedSwath/gap=5000000/full_scan-10                  633.3Ki ± ∞ ¹   569.5Ki ± ∞ ¹  -10.08% (p=0.029 n=4)
PointDeletedSwath/gap=10000000/prefix_point_lookup-10         77.00 ± ∞ ¹     48.00 ± ∞ ¹  -37.66% (p=0.029 n=4)
PointDeletedSwath/gap=10000000/non-prefix_point_seek-10   747232.00 ± ∞ ¹     48.00 ± ∞ ¹  -99.99% (p=0.029 n=4)
PointDeletedSwath/gap=10000000/full_scan-10                730.48Ki ± ∞ ¹   49.18Ki ± ∞ ¹  -93.27% (p=0.029 n=4)
geomean                                                     2.953Ki         1.352Ki        -54.22%
¹ need >= 6 samples for confidence interval at level 0.95
                                                        │     swath      │              swath-heur               │
                                                        │   allocs/op    │  allocs/op    vs base                 │
PointDeletedSwath/gap=100/prefix_point_lookup-10             4.000 ± ∞ ¹    4.000 ± ∞ ¹        ~ (p=1.000 n=4) ²
PointDeletedSwath/gap=100/non-prefix_point_seek-10           4.000 ± ∞ ¹    4.000 ± ∞ ¹        ~ (p=1.000 n=4) ²
PointDeletedSwath/gap=100/full_scan-10                      6.684k ± ∞ ¹   8.862k ± ∞ ¹  +32.60% (p=0.029 n=4)
PointDeletedSwath/gap=1000/prefix_point_lookup-10            4.000 ± ∞ ¹    4.000 ± ∞ ¹        ~ (p=1.000 n=4) ²
PointDeletedSwath/gap=1000/non-prefix_point_seek-10          4.000 ± ∞ ¹    4.000 ± ∞ ¹        ~ (p=1.000 n=4) ²
PointDeletedSwath/gap=1000/full_scan-10                     6.700k ± ∞ ¹   8.844k ± ∞ ¹  +31.99% (p=0.029 n=4)
PointDeletedSwath/gap=10000/prefix_point_lookup-10           4.000 ± ∞ ¹    4.000 ± ∞ ¹        ~ (p=1.000 n=4) ²
PointDeletedSwath/gap=10000/non-prefix_point_seek-10         4.000 ± ∞ ¹    4.000 ± ∞ ¹        ~ (p=1.000 n=4) ²
PointDeletedSwath/gap=10000/full_scan-10                    6.711k ± ∞ ¹   8.789k ± ∞ ¹  +30.96% (p=0.029 n=4)
PointDeletedSwath/gap=100000/prefix_point_lookup-10          4.000 ± ∞ ¹    4.000 ± ∞ ¹        ~ (p=1.000 n=4) ²
PointDeletedSwath/gap=100000/non-prefix_point_seek-10        4.000 ± ∞ ¹    4.000 ± ∞ ¹        ~ (p=1.000 n=4) ²
PointDeletedSwath/gap=100000/full_scan-10                   6.750k ± ∞ ¹   8.784k ± ∞ ¹  +30.12% (p=0.029 n=4)
PointDeletedSwath/gap=200000/prefix_point_lookup-10          4.000 ± ∞ ¹    4.000 ± ∞ ¹        ~ (p=1.000 n=4) ²
PointDeletedSwath/gap=200000/non-prefix_point_seek-10        4.000 ± ∞ ¹    4.000 ± ∞ ¹        ~ (p=1.000 n=4) ²
PointDeletedSwath/gap=200000/full_scan-10                   6.827k ± ∞ ¹   8.704k ± ∞ ¹  +27.51% (p=0.029 n=4)
PointDeletedSwath/gap=400000/prefix_point_lookup-10          4.000 ± ∞ ¹    4.000 ± ∞ ¹        ~ (p=1.000 n=4) ²
PointDeletedSwath/gap=400000/non-prefix_point_seek-10        5.000 ± ∞ ¹    4.000 ± ∞ ¹  -20.00% (p=0.029 n=4)
PointDeletedSwath/gap=400000/full_scan-10                   6.979k ± ∞ ¹   8.626k ± ∞ ¹  +23.61% (p=0.029 n=4)
PointDeletedSwath/gap=5000000/prefix_point_lookup-10         4.000 ± ∞ ¹    3.000 ± ∞ ¹  -25.00% (p=0.029 n=4)
PointDeletedSwath/gap=5000000/non-prefix_point_seek-10      95.000 ± ∞ ¹    3.000 ± ∞ ¹  -96.84% (p=0.029 n=4)
PointDeletedSwath/gap=5000000/full_scan-10                  6.809k ± ∞ ¹   6.136k ± ∞ ¹   -9.87% (p=0.029 n=4)
PointDeletedSwath/gap=10000000/prefix_point_lookup-10        4.000 ± ∞ ¹    3.000 ± ∞ ¹  -25.00% (p=0.029 n=4)
PointDeletedSwath/gap=10000000/non-prefix_point_seek-10   7589.000 ± ∞ ¹    3.000 ± ∞ ¹  -99.96% (p=0.029 n=4)
PointDeletedSwath/gap=10000000/full_scan-10                 7664.5 ± ∞ ¹    519.0 ± ∞ ¹  -93.23% (p=0.029 n=4)
geomean                                                      75.61          43.37        -42.64%
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
</pre>
</details>

Seek/scan performance is greatly improved in all cases, with the most extreme example being a non-prefix `SeekGE` which went from 1.17 sec -> 4 microseconds.

### `BenchmarkQueueWorkload`
<details>
<summary>Show Results</summary>
<pre>
goos: darwin
goarch: arm64
pkg: github.com/cockroachdb/pebble
                                                                           │       queue       │               queue-heur                │
                                                                           │      sec/op       │     sec/op      vs base                 │
QueueWorkload/initial_ops=400000/deleteRatio=0.10/valueSize=128/seek-10         1259.68µ ± ∞ ¹     30.23µ ± ∞ ¹    -97.60% (p=0.029 n=4)
QueueWorkload/initial_ops=800000/deleteRatio=0.10/valueSize=128/seek-10         2513.78µ ± ∞ ¹     31.15µ ± ∞ ¹    -98.76% (p=0.029 n=4)
QueueWorkload/initial_ops=1200000/deleteRatio=0.10/valueSize=128/seek-10        3731.32µ ± ∞ ¹     32.12µ ± ∞ ¹    -99.14% (p=0.029 n=4)
QueueWorkload/initial_ops=2000000/deleteRatio=0.10/valueSize=128/seek-10          19.73µ ± ∞ ¹     22.73µ ± ∞ ¹    +15.20% (p=0.029 n=4)
QueueWorkload/initial_ops=3500000/deleteRatio=0.10/valueSize=128/seek-10       15736.10µ ± ∞ ¹     32.76µ ± ∞ ¹    -99.79% (p=0.029 n=4)
QueueWorkload/initial_ops=5000000/deleteRatio=0.10/valueSize=128/seek-10          58.04µ ± ∞ ¹    102.67µ ± ∞ ¹    +76.89% (p=0.029 n=4)
QueueWorkload/initial_ops=7500000/deleteRatio=0.10/valueSize=128/seek-10        36274.4µ ± ∞ ¹     129.6µ ± ∞ ¹    -99.64% (p=0.029 n=4)
QueueWorkload/initial_ops=10000000/deleteRatio=0.10/valueSize=128/seek-10       90382.4µ ± ∞ ¹     155.2µ ± ∞ ¹    -99.83% (p=0.029 n=4)
QueueWorkload/initial_ops=50000000/deleteRatio=0.10/valueSize=128/seek-10      419112.4µ ± ∞ ¹     793.9µ ± ∞ ¹    -99.81% (p=0.029 n=4)
QueueWorkload/initial_ops=400000/deleteRatio=0.10/valueSize=2048/seek-10          1.420m ± ∞ ¹     1.391m ± ∞ ¹          ~ (p=0.343 n=4)
QueueWorkload/initial_ops=800000/deleteRatio=0.10/valueSize=2048/seek-10          39.38m ± ∞ ¹     21.81m ± ∞ ¹    -44.62% (p=0.029 n=4)
QueueWorkload/initial_ops=1200000/deleteRatio=0.10/valueSize=2048/seek-10      63213.88µ ± ∞ ¹     69.63µ ± ∞ ¹    -99.89% (p=0.029 n=4)
QueueWorkload/initial_ops=2000000/deleteRatio=0.10/valueSize=2048/seek-10        75.239m ± ∞ ¹     6.615m ± ∞ ¹    -91.21% (p=0.029 n=4)
QueueWorkload/initial_ops=3500000/deleteRatio=0.10/valueSize=2048/seek-10        102.56m ± ∞ ¹     10.33m ± ∞ ¹    -89.93% (p=0.029 n=4)
QueueWorkload/initial_ops=5000000/deleteRatio=0.10/valueSize=2048/seek-10     193040.49µ ± ∞ ¹     97.56µ ± ∞ ¹    -99.95% (p=0.029 n=4)
QueueWorkload/initial_ops=7500000/deleteRatio=0.10/valueSize=2048/seek-10       450.019m ± ∞ ¹     2.403m ± ∞ ¹    -99.47% (p=0.029 n=4)
QueueWorkload/initial_ops=10000000/deleteRatio=0.10/valueSize=2048/seek-10       633.85m ± ∞ ¹     21.44m ± ∞ ¹    -96.62% (p=0.029 n=4)
QueueWorkload/initial_ops=50000000/deleteRatio=0.10/valueSize=2048/seek-10    3410138.2µ ± ∞ ¹     190.3µ ± ∞ ¹    -99.99% (p=0.029 n=4)
QueueWorkload/initial_ops=400000/deleteRatio=0.30/valueSize=128/seek-10         3789.61µ ± ∞ ¹     19.74µ ± ∞ ¹    -99.48% (p=0.029 n=4)
QueueWorkload/initial_ops=800000/deleteRatio=0.30/valueSize=128/seek-10         7475.93µ ± ∞ ¹     19.91µ ± ∞ ¹    -99.73% (p=0.029 n=4)
QueueWorkload/initial_ops=1200000/deleteRatio=0.30/valueSize=128/seek-10       11237.05µ ± ∞ ¹     30.07µ ± ∞ ¹    -99.73% (p=0.029 n=4)
QueueWorkload/initial_ops=2000000/deleteRatio=0.30/valueSize=128/seek-10          19.89µ ± ∞ ¹     20.85µ ± ∞ ¹          ~ (p=0.343 n=4)
QueueWorkload/initial_ops=3500000/deleteRatio=0.30/valueSize=128/seek-10       11091.99µ ± ∞ ¹     20.15µ ± ∞ ¹    -99.82% (p=0.029 n=4)
QueueWorkload/initial_ops=5000000/deleteRatio=0.30/valueSize=128/seek-10         827.15µ ± ∞ ¹     58.48µ ± ∞ ¹    -92.93% (p=0.029 n=4)
QueueWorkload/initial_ops=7500000/deleteRatio=0.30/valueSize=128/seek-10       22341.01µ ± ∞ ¹     69.57µ ± ∞ ¹    -99.69% (p=0.029 n=4)
QueueWorkload/initial_ops=10000000/deleteRatio=0.30/valueSize=128/seek-10     195604.48µ ± ∞ ¹     59.07µ ± ∞ ¹    -99.97% (p=0.029 n=4)
QueueWorkload/initial_ops=50000000/deleteRatio=0.30/valueSize=128/seek-10      282298.4µ ± ∞ ¹     148.3µ ± ∞ ¹    -99.95% (p=0.029 n=4)
QueueWorkload/initial_ops=400000/deleteRatio=0.30/valueSize=2048/seek-10          61.49µ ± ∞ ¹    61.32µ ± ∞ ¹         ~ (p=0.200 n=4)
QueueWorkload/initial_ops=800000/deleteRatio=0.30/valueSize=2048/seek-10          61.50µ ± ∞ ¹    117.69µ ± ∞ ¹    +91.35% (p=0.029 n=4)
QueueWorkload/initial_ops=1200000/deleteRatio=0.30/valueSize=2048/seek-10        111.72m ± ∞ ¹     11.06m ± ∞ ¹    -90.10% (p=0.029 n=4)
QueueWorkload/initial_ops=2000000/deleteRatio=0.30/valueSize=2048/seek-10     155366.61µ ± ∞ ¹     72.87µ ± ∞ ¹    -99.95% (p=0.029 n=4)
QueueWorkload/initial_ops=3500000/deleteRatio=0.30/valueSize=2048/seek-10     211515.99µ ± ∞ ¹     73.99µ ± ∞ ¹    -99.97% (p=0.029 n=4)
QueueWorkload/initial_ops=5000000/deleteRatio=0.30/valueSize=2048/seek-10     389765.22µ ± ∞ ¹     73.84µ ± ∞ ¹    -99.98% (p=0.029 n=4)
QueueWorkload/initial_ops=7500000/deleteRatio=0.30/valueSize=2048/seek-10      525141.9µ ± ∞ ¹     116.3µ ± ∞ ¹    -99.98% (p=0.029 n=4)
QueueWorkload/initial_ops=10000000/deleteRatio=0.30/valueSize=2048/seek-10   1519714.54µ ± ∞ ¹     65.99µ ± ∞ ¹   -100.00% (p=0.029 n=4)
geomean                                                                           21.40m           233.7µ          -98.91%
¹ need >= 6 samples for confidence interval at level 0.95
                                                                           │     queue      │               queue-heur                │
                                                                           │      B/op      │     B/op       vs base                  │
QueueWorkload/initial_ops=400000/deleteRatio=0.10/valueSize=128/seek-10         196.0 ± ∞ ¹     384.0 ± ∞ ¹   +95.92% (p=0.029 n=4)
QueueWorkload/initial_ops=800000/deleteRatio=0.10/valueSize=128/seek-10         233.5 ± ∞ ¹     384.0 ± ∞ ¹   +64.45% (p=0.029 n=4)
QueueWorkload/initial_ops=1200000/deleteRatio=0.10/valueSize=128/seek-10        255.0 ± ∞ ¹     384.0 ± ∞ ¹   +50.59% (p=0.029 n=4)
QueueWorkload/initial_ops=2000000/deleteRatio=0.10/valueSize=128/seek-10        192.0 ± ∞ ¹     192.0 ± ∞ ¹         ~ (p=1.000 n=4) ²
QueueWorkload/initial_ops=3500000/deleteRatio=0.10/valueSize=128/seek-10        576.5 ± ∞ ¹     384.0 ± ∞ ¹   -33.39% (p=0.029 n=4)
QueueWorkload/initial_ops=5000000/deleteRatio=0.10/valueSize=128/seek-10        193.0 ± ∞ ¹     386.0 ± ∞ ¹  +100.00% (p=0.029 n=4)
QueueWorkload/initial_ops=7500000/deleteRatio=0.10/valueSize=128/seek-10       1708.0 ± ∞ ¹     385.0 ± ∞ ¹   -77.46% (p=0.029 n=4)
QueueWorkload/initial_ops=10000000/deleteRatio=0.10/valueSize=128/seek-10      2937.0 ± ∞ ¹     577.0 ± ∞ ¹   -80.35% (p=0.029 n=4)
QueueWorkload/initial_ops=50000000/deleteRatio=0.10/valueSize=128/seek-10     10330.0 ± ∞ ¹     592.0 ± ∞ ¹   -94.27% (p=0.029 n=4)
QueueWorkload/initial_ops=400000/deleteRatio=0.10/valueSize=2048/seek-10        409.5 ± ∞ ¹     409.0 ± ∞ ¹         ~ (p=0.486 n=4)
QueueWorkload/initial_ops=800000/deleteRatio=0.10/valueSize=2048/seek-10       1566.0 ± ∞ ¹    1013.0 ± ∞ ¹         ~ (p=0.057 n=4)
QueueWorkload/initial_ops=1200000/deleteRatio=0.10/valueSize=2048/seek-10      2058.5 ± ∞ ¹     192.5 ± ∞ ¹   -90.65% (p=0.029 n=4)
QueueWorkload/initial_ops=2000000/deleteRatio=0.10/valueSize=2048/seek-10      1931.5 ± ∞ ¹     526.0 ± ∞ ¹   -72.77% (p=0.029 n=4)
QueueWorkload/initial_ops=3500000/deleteRatio=0.10/valueSize=2048/seek-10      3796.0 ± ∞ ¹     955.0 ± ∞ ¹   -74.84% (p=0.029 n=4)
QueueWorkload/initial_ops=5000000/deleteRatio=0.10/valueSize=2048/seek-10      4317.5 ± ∞ ¹     385.0 ± ∞ ¹   -91.08% (p=0.029 n=4)
QueueWorkload/initial_ops=7500000/deleteRatio=0.10/valueSize=2048/seek-10     10816.0 ± ∞ ¹     684.0 ± ∞ ¹   -93.68% (p=0.029 n=4)
QueueWorkload/initial_ops=10000000/deleteRatio=0.10/valueSize=2048/seek-10   15.980Ki ± ∞ ¹   1.258Ki ± ∞ ¹   -92.13% (p=0.029 n=4)
QueueWorkload/initial_ops=50000000/deleteRatio=0.10/valueSize=2048/seek-10    36896.0 ± ∞ ¹     579.0 ± ∞ ¹   -98.43% (p=0.029 n=4)
QueueWorkload/initial_ops=400000/deleteRatio=0.30/valueSize=128/seek-10         255.0 ± ∞ ¹     192.0 ± ∞ ¹   -24.71% (p=0.029 n=4)
QueueWorkload/initial_ops=800000/deleteRatio=0.30/valueSize=128/seek-10         319.0 ± ∞ ¹     192.0 ± ∞ ¹   -39.81% (p=0.029 n=4)
QueueWorkload/initial_ops=1200000/deleteRatio=0.30/valueSize=128/seek-10        393.0 ± ∞ ¹     384.0 ± ∞ ¹         ~ (p=0.143 n=4)
QueueWorkload/initial_ops=2000000/deleteRatio=0.30/valueSize=128/seek-10        192.0 ± ∞ ¹     192.0 ± ∞ ¹         ~ (p=1.000 n=4) ²
QueueWorkload/initial_ops=3500000/deleteRatio=0.30/valueSize=128/seek-10        603.0 ± ∞ ¹     192.0 ± ∞ ¹   -68.16% (p=0.029 n=4)
QueueWorkload/initial_ops=5000000/deleteRatio=0.30/valueSize=128/seek-10        393.5 ± ∞ ¹     193.0 ± ∞ ¹   -50.95% (p=0.029 n=4)
QueueWorkload/initial_ops=7500000/deleteRatio=0.30/valueSize=128/seek-10       1315.5 ± ∞ ¹     385.0 ± ∞ ¹   -70.73% (p=0.029 n=4)
QueueWorkload/initial_ops=10000000/deleteRatio=0.30/valueSize=128/seek-10      5442.0 ± ∞ ¹     193.0 ± ∞ ¹   -96.45% (p=0.029 n=4)
QueueWorkload/initial_ops=50000000/deleteRatio=0.30/valueSize=128/seek-10      5272.0 ± ∞ ¹     578.0 ± ∞ ¹   -89.04% (p=0.029 n=4)
QueueWorkload/initial_ops=400000/deleteRatio=0.30/valueSize=2048/seek-10        192.5 ± ∞ ¹     540.0 ± ∞ ¹  +180.52% (p=0.029 n=4)
QueueWorkload/initial_ops=800000/deleteRatio=0.30/valueSize=2048/seek-10        193.0 ± ∞ ¹     385.5 ± ∞ ¹   +99.74% (p=0.029 n=4)
QueueWorkload/initial_ops=1200000/deleteRatio=0.30/valueSize=2048/seek-10      3438.0 ± ∞ ¹     719.0 ± ∞ ¹   -79.09% (p=0.029 n=4)
QueueWorkload/initial_ops=2000000/deleteRatio=0.30/valueSize=2048/seek-10      4431.5 ± ∞ ¹     193.0 ± ∞ ¹   -95.64% (p=0.029 n=4)
QueueWorkload/initial_ops=3500000/deleteRatio=0.30/valueSize=2048/seek-10      6739.0 ± ∞ ¹     192.5 ± ∞ ¹   -97.14% (p=0.029 n=4)
…